+++ helm-charts/strimzi-kafka-operator/templates/044-Crd-kafkauser.yaml
@@ -138,9 +138,6 @@ spec:
               type: integer
             username:
               type: string
-            credentials:
-              type: object
-              properties:
-                secret:
-                  type: string
+            secret:
+              type: string
 {{- end -}}
